Added ability to change viewpoint "on the fly" with hot keys
Refactored to allow for greater control over initialization and
re-use of objects, and scheduled a new thread to watch for keyboard
input. Note that we can't read directly from System.in() because
that is a blocking method.

"f" switches to FPV
"g" switches to Gimbal view
"s" switches to the static camera on the ground